+++ CHANGELOG.md
@@ -13,11 +13,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - PMD for static source code analysis
 - *Experimental* GraalVM support. Run keycloak-config-cli without Java!
+- Throw errors on unknown properties in config files
 
 ### Changed
 - Reduce docker image size
 - Bump SpringBoot from 2.2.7 to 2.3.0
-- Bump spotbugs from 4.0.1 to 4.0.2
 - Bump keycloak from 10.0.0 to 10.0.1
 - Used keycloak parent pom instead manage versions of 3rd party libs
 - Add experimental profile for spring native builds